refactor provider
