+++ examples/providers/src/presentation.rs
@@ -7,8 +7,8 @@ use chrono::{DateTime, Utc};
 use ecdsa::signature::Signer as _;
 use ecdsa::{Signature, SigningKey};
 use k256::Secp256k1;
-use vercre_core::callback::Payload;
-use vercre_core::metadata::{self, VpFormat};
+use vercre_core::provider::Payload;
+use vercre_core::provider::{self as metadata, VpFormat};
 use vercre_core::provider::{Callback, ClientMetadata, Result, StateManager};
 use vercre_core::vci::{Format, GrantType};
 use vercre_vc::proof::{Algorithm, Signer};
