+++ vercre-holder/src/credential.rs
@@ -6,7 +6,7 @@
 use std::ops::Deref;
 
 use serde::{Deserialize, Serialize};
-use vercre_core::metadata::CredentialConfiguration;
+use vercre_core::provider::CredentialConfiguration;
 use vercre_vc::model::vc::VerifiableCredential;
 
 /// The Credential model contains information about a credential owned by the
