+++ vercre-holder/src/presentation.rs
@@ -365,7 +365,7 @@ mod tests {
 
     use insta::assert_yaml_snapshot as assert_snapshot;
     use providers::{issuance, presentation};
-    use vercre_core::metadata::CredentialConfiguration;
+    use vercre_core::provider::CredentialConfiguration;
     use vercre_exch::{
         Field, Filter, FilterValue, Format, InputDescriptor, PresentationDefinition,
     };
