+++ vercre-holder/src/provider.rs
@@ -14,8 +14,8 @@ pub use client::{IssuerClient, VerifierClient};
 pub use input::{IssuanceInput, PresentationInput};
 pub use status_listener::{IssuanceListener, PresentationListener};
 pub use storer::CredentialStorer;
-pub use vercre_core::metadata::CredentialConfiguration;
+pub use vercre_core::provider::CredentialConfiguration;
 pub use vercre_core::provider::{Callback, Result, StateManager};
-pub use vercre_core::signature::{Algorithm, Signer};
+pub use vercre_core::provider::{Algorithm, Signer};
 pub use vercre_core::vci::TxCode;
 pub use vercre_exch::Constraints;
