+++ vercre-issuer/src/batch.rs
@@ -14,7 +14,7 @@ use anyhow::anyhow;
 use chrono::Utc;
 use tracing::instrument;
 use vercre_core::error::{Ancillary as _, Err};
-use vercre_core::metadata::{CredentialDefinition, Issuer};
+use vercre_core::provider::{CredentialDefinition, Issuer};
 use vercre_core::provider::{
     Callback, ClientMetadata, IssuerMetadata, ServerMetadata, StateManager, Subject,
 };
