+++ vercre-issuer/src/lib.rs
@@ -104,7 +104,7 @@ use vercre_core::provider::{
 };
 // TODO: move Claims into jwt module
 pub use vercre_core::vci::{GrantType, ProofClaims};
-pub use vercre_core::{callback, provider, subject, Result};
+pub use vercre_core::{ provider, Result};
 use vercre_vc::proof::Signer;
 
 /// Endpoint is used to surface the public Verifiable Presentation endpoints to
