+++ vercre-verifier/src/metadata.rs
@@ -8,7 +8,7 @@
 use std::fmt::Debug;
 
 use tracing::instrument;
-pub use vercre_core::metadata as types;
+pub use vercre_core::provider as types;
 use vercre_core::provider::{Callback, ClientMetadata, StateManager};
 #[allow(clippy::module_name_repetitions)]
 pub use vercre_core::vp::{MetadataRequest, MetadataResponse};
