+++ vercre-verifier/src/request.rs
@@ -18,8 +18,7 @@ use std::fmt::Debug;
 use anyhow::anyhow;
 use tracing::instrument;
 use vercre_core::error::Err;
-use vercre_core::provider::{Callback, ClientMetadata, StateManager};
-use vercre_core::signature::Signer;
+use vercre_core::provider::{Callback, ClientMetadata, Signer, StateManager};
 #[allow(clippy::module_name_repetitions)]
 pub use vercre_core::vp::{RequestObjectRequest, RequestObjectResponse};
 use vercre_core::{err, jws, Result};
